fix: change total type to long in MoMo payment request

+++ frontend/app/src/main/java/com/example/my_mobile_app/ui/checkout/CheckoutActivity.java
@@ -711,7 +711,7 @@ public class CheckoutActivity extends BaseActivity {
     }
 
     private void createMomoPayment(Order order) {
-        double total = subtotal + shippingFee;
+        long total = (long)(subtotal + shippingFee);
         CreateMoMoPaymentRequest req = new CreateMoMoPaymentRequest(
                 order.orderId, total, getString(R.string.checkout_momo_order_info, order.orderId));
         req.requestType = "captureWallet";
